refactor: move issues to actual tests and refactor
